+++ HWs/HW2/module/Makefile
@@ -1,4 +1,4 @@
-obj-m += hw1secws.o
+obj-m += hw2secws.o
 
 all:
 	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
