--- HWs/HW2_COPY/module/Makefile
@@ -1,7 +1,0 @@
-obj-m += hw2secws.o
-
-all:
-	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) modules
-
-clean:
-	make -C /lib/modules/$(shell uname -r)/build M=$(PWD) clean
